requirements.txt
Flask~=2.0.1
pyzbar~=0.1.8
Pillow~=8.2.0
loguru~=0.5.3
pytesseract~=0.3.8
pyzbar~=0.1.8
